requirements.txt
transformers
datasets
huggingface_hub
torch-lr-finder
python-dotenv
pandas
tqdm